requirements.txt
streamlit
Pillow==8.3.2
tensorflow-cpu
tensorflow
numpy
pytest-shutil